added a runscript for models.
The script has the following usage:
	run_actr.sh model.pl
where model.pl is a CHR-ACT-R model.
Basically, the script defines a new search path called chractr
which links to the location of the script (i.e. where the CHR-ACT-R
environment can be found in the file system).
Models which cooperate with the script then do not include the core
relatively to their location anymore, but use the following include line:
  :- include(chractr('actr_core.pl').
allowing them to be located in different directories than the core.
For legacy reasons, they add '.' to the file search path, so they can
be loaded as in previous versions (by simply consulting the model which
must then be located in the same directory as the core).

The compiler will be adapted for the new usage.